Expose RTCIceCandidate and RTCSessionDescription
This places the API in line with what node addons like wrtc provide, where
the existence of these functions is not guaranteed. As such, this change
should make application code which must run both server- and client-side
slightly easier to program by guaranteeing that the functions are included
with the module.