Fully disable all console logging by default
Within the `utils` module, there is a logging statement (which logs
whether the browser is below the minimum supported version). As such, we
have to properly initialize our logging as disabled _within_ the utils
module, rather than overwriting the logging setting as disabled only in
the core module. We do not want console logging to be on anywhere by
default, since it is considered bad practice in production code.