edge: add partial support for unified plan
adds partial support for Unified Plan in Edge.
This is limited to creating offers after calling addStream multiple times.
Renegotiation is not supported.

On the receiving end this adds support for parsing MSID and overrides
the native stream and track ids with the stream and track ids from the
SDP.