The shim for webOS 3.0.0+ fixing data-channel
Hi,

I want to propose shim for webOS 3.x, which fixes problems with data-channels. Because SmartTV uses Chromium v38 based on blink. So I decided to make updates in the shim of Chrome browser.

Please, have a look and contact me if there any questions on styling or placing.
Thanks, Sergey.