ensure shim’s name property matches original function (#978)
Description
Adds function names to shim functions at point of definition.

Purpose
This is just a small change to make the shims act a little more like the original functions. As is if you call RTCIceCandidate.name on an unshimmed RTCIceCandidate, you get the string "RTCIceCandidate" but on a shimmed one you get "".